added default app.tsx
added app.tsx as default gateway
