+++ frontend/app/App.tsx
@@ -1,0 +1,7 @@
+
+import React from "react";
+import Home from "./screens/Home";
+
+export default function App() {
+  return <Home />;
+}
