+++ frontend/package.json
@@ -27,7 +27,8 @@
     "react-native-safe-area-context": "4.10.1",
     "react-native-screens": "3.31.1",
     "react-native-snap-carousel": "^1.6.1",
-    "twrnc": "^4.2.0"
+    "twrnc": "^4.2.0",
+    "expo-navigation-bar": "~3.0.5"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
